converting from javascript to typescript
