+++ apps/web/pages/_app.tsx
@@ -1,0 +1,6 @@
+import './styles.css'
+import type { AppProps } from 'next/app'
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />
+}
